requirements.txt
fastapi
uvicorn
langchain
openai
faiss-cpu
PyPDF2
python-dotenv
streamlit 
